Restructure project layout
This is a big change. It adds a common top-level directory.
The common directory has rake tasks that can be used to create
the binary input files.
It is a top-level directory so it can refer to configuration
file paths and have them resolve when used by the gems.

It also moves the api-common-protos and protobuf submodules
to the common directory.

The input files and expected output files are shared via
symlinks. In the future, new gems can replace their local
directories with these symlinks to use the shared files.

The speech binary file has been updated with the latest protos,
and the vision binary file has also been added.

The executable binary file has been refactored to read from
STDIN and write to STDOUT. It creates the CodeGeneratorRequest
object and passes it to the runner. Customized gems now pass
the preferred generator type to the runner.

Use of environment variables have been removed in favor of
using values specified by the protoc options arguments.