Refactor base generator code formatting
Move the formatter into the base generator.
Rename the generate_file method, add convenience name as alias.
Add config method that allows plugin generators to override
the location of the configuration file to use to format.
Add formatter code as separate method so it can also be
overridden. (Hi, rubyfmt)
Fix the formatter so it works in the cloud generator.