Replace implicit kwarg hash passing with the `**` operator.
In Rubies below 2.7, passing kwargs as the only hash argument is a
supported but unintentional valid way to call a function. However, in
[Ruby 2.7](https://www.ruby-lang.org/en/news/2019/12/12/separation-of-positional-and-keyword-arguments-in-ruby-3-0/)
this is a warning, and will break in Ruby 3.0.

This commit replaces all instances of passing hashes in kwarg positions with the
`**` operator, which is valid in old rubies, as well as 2.7.