Preserve call options in LROs
With ads, customer ID and Developer Token are passed as call metadata,
as such, we need them to be preserved to follow on calls made through
call metadata.

This patch ensures that `Gapic::Operation` instances created in clients
that use an `OperationsClient` preserve metadata from the calls that
created them.